+++ pages/FhirMapper2/index.vue
@@ -944,7 +944,7 @@ group SetEntryData(source src: Patient, target entry)
       // brianpos hosted service
       // default the firely SDK/brianpos service
       // Source code for this is at https://github.com/brianpos/fhirpath-lab-dotnet
-      let url = settings.getServerEngineUrl("mapper_server");
+      let url = (await settings.getServerEngineUrl("mapper_server"));
 
       let p: fhir4b.Parameters = {
         resourceType: "Parameters",
@@ -963,7 +963,7 @@ group SetEntryData(source src: Patient, target entry)
       }
 
       if (this.selectedEngine == "java (HAPI)") {
-        url = settings.getServerEngineUrl("mapper_server_java");
+        url = (await settings.getServerEngineUrl("mapper_server_java"));
 
         if (!this.getResourceJson() && this.resourceId) {
           await this.downloadTestResource();