+++ pages/Questionnaire/tester.vue
@@ -2557,7 +2557,7 @@ export default Vue.extend({
           url.startsWith("https://build.fhir.org/") ||
           url.startsWith("https://hl7.org/fhir/")
         )
-          url = settings.dotnet_server_downloader() + "?url=" + url;
+          url = (await settings.dotnet_server_downloader()) + "?url=" + url;
 
         if (this.cancelSource) this.cancelSource.cancel("new download started");
         this.cancelSource = axios.CancelToken.source();
@@ -2667,7 +2667,7 @@ export default Vue.extend({
           url.startsWith("https://build.fhir.org/") ||
           url.startsWith("https://hl7.org/fhir/")
         )
-          url = settings.dotnet_server_downloader() + "?url=" + url;
+          url = (await settings.dotnet_server_downloader()) + "?url=" + url;
 
         if (this.cancelSource) this.cancelSource.cancel("new download started");
         this.cancelSource = axios.CancelToken.source();