+++ pages/fml.vue
@@ -1698,7 +1698,7 @@ group SetEntryData(source src: Patient, target entry)
       // brianpos hosted service
       // default the firely SDK/brianpos service
       // Source code for this is at https://github.com/brianpos/fhirpath-lab-dotnet
-      let url = settings.getServerEngineUrl("mapper_server");
+      let url = (await settings.getServerEngineUrl("mapper_server"));
 
       let p: fhir4b.Parameters = {
         resourceType: "Parameters",
@@ -1720,11 +1720,11 @@ group SetEntryData(source src: Patient, target entry)
       }
 
       if (this.selectedEngine == "java (HAPI)") {
-        url = settings.getServerEngineUrl("mapper_server_java");
+        url = (await settings.getServerEngineUrl("mapper_server_java"));
         (this as any).$appInsights?.trackEvent({ name: 'evaluate HAPI (map)' });
       }
       else if (this.selectedEngine == "matchbox") {
-        url = settings.getServerEngineUrl("mapper_server_matchbox");
+        url = (await settings.getServerEngineUrl("mapper_server_matchbox"));
         (this as any).$appInsights?.trackEvent({ name: 'evaluate matchbox (map)' });
       }
       else {